refact: extract folders to a dedicated file
