--- src-tauri/src/errors/mod.rs
@@ -1,6 +1,0 @@
-pub mod controller;
-
-
-
-
-
