--- src-tauri/src/utils/aplication_state.rs
@@ -1,5 +1,0 @@
-use rocket::Shutdown;
-
-#[derive(Default)]
-pub struct RocketShutdownHandle(pub Option<Shutdown>);
-
