--- src-tauri/src/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod local_ip;
-pub mod aplication_state;
-
